client/requirements.txt
## non python requirements
# ipfs version 0.14.0 - 0.17.0

setuptools==65.6.3 # 56.0.0
# needed for engine
xgboost==1.7.2 # 1.5.1
pandas==1.5.2 # 1.3.4
joblib==1.2.0 # 1.0.1
numpy==1.24.0 # 1.23.0
ppscore==1.3.0 # 1.2.0
scikit-learn==1.2.0 # 0.24.2
sklearn==0.0.post1 # 0.0

# needed for data managemnet
pyarrow==10.0.1 # 6.0.1
fastparquet==2022.12.0 # 0.7.2 # fastparquet-2022.12.0

# needed for configuration management
PyYAML==6.0 #5.4.1

# needed for node architecture
reactivex==4.0.4 #4.0.0b5

# needed for getting system data
psutil==5.9.0 #5.8.0

# needed for local webserver (UI)
waitress==2.1.2 #2.1.1
Flask==2.2.2 # 2.1.1
WTForms==3.0.1
flask_wtf==1.0.1

# needed for wallet
python-ravencoinlib==0.2.2
mnemonic==0.20
base58==2.1.1

# needed for cli
click==8.1.3 #8.1.2

# needed for signing messages for connection with Satori server
Crypto==1.4.1

# needed for Satori pubsub connection
websocket-client==1.4.2 # not sure if we need this
websockets==10.4 # import websocket; websocket.WebSocket()

## needed? for old view
#IPython==8.7.0  # 7.23.0
#matplotlib==3.6.2 # 3.4.1
#seaborn==0.11.2
## needed? for demo data
#forex-python==1.8
## needed? for google sheets
#oauth2client==4.1.3
#gspread==5.7.2 #5.0.0

## attempt to avoid error on macos about unsafe use of cryptolib
#pyOpenSSL==22.0.0
#cryptography==37.0.4